Add better testing
